@@ -83,7 +83,7 @@ private:
     // ======================================================================
     auto &value_as_string()
     {
-        return boost::get<string_value>(value());
+        return std::get<string_value>(value());
     }
 
     // ======================================================================
@@ -91,7 +91,7 @@ private:
     // ======================================================================
     auto &value_as_array()
     {
-        return boost::get<array_value>(value());
+        return std::get<array_value>(value());
     }
 
     // ======================================================================
@@ -99,7 +99,7 @@ private:
     // ======================================================================
     auto &value_as_table()
     {
-        return boost::get<table_value>(value());
+        return std::get<table_value>(value());
     }
 
     // ======================================================================
